build: zero-config defaults + warning-clean MSVC /W3 /sdl /WX
Default macros (no -D needed to consume the amalgam):
- Linkage: static is the default (HOOX_API empty). Opt into a Windows DLL with
  HOOX_SHARED (+ HOOX_EXPORTS when building it). Was: dllimport unless
  HOOX_STATIC.
- Allocator: system malloc is the default. Opt into dlmalloc with
  HOOX_USE_DLMALLOC. Was: dlmalloc unless HOOX_USE_SYSTEM_ALLOC (and dlmalloc.c
  isn't vendored, so the old default didn't even build).
- Arch/OS: hooxdefs.h derives HAVE_I386/HAVE_WINDOWS/... from compiler built-ins
  (_M_IX86/__x86_64__/_WIN32/...); user-defined only as an override. CMake still
  passes them explicitly as the authoritative cross-compile override.
- example + test_amalgam now build with no feature macros at all.

MSVC strictness (request: /sdl + warnings-as-errors, all warnings fixed):
- Top-level CMake adds /W3 /GF /sdl /WX for MSVC.
- Fixed every resulting warning: 34 narrowing conversions (size_t/int64 ->
  smaller ints) made explicit across writer/relocator/codeallocator/tls/memory/
  interceptor/cpu + the arch-x86 relocator test.
- HX_GNUC_INTERNAL is empty on _WIN32 (PE has no ELF visibility; gcc warned).

Verified: warning-clean full build + suite on MSVC 19.38 (/W3 /sdl /WX),
clang 19 and gcc 13; example builds+runs under all three, and under MSVC
/W3 /sdl /WX with zero feature macros.

+++ src/backend/windows/hooxtls-windows.c
@@ -52,7 +52,7 @@ hoox_tls_key_new (void)
 void
 hoox_tls_key_free (HooxTlsKey key)
 {
-  TlsFree (key);
+  TlsFree ((DWORD) key);
 }
 
 void
@@ -211,7 +211,7 @@ hoox_tls_key_get_value (HooxTlsKey key)
 {
   if (key < 64)
   {
-    return (hx_pointer) __readgsqword (0x1480 + key * sizeof (hx_pointer));
+    return (hx_pointer) __readgsqword ((DWORD) (0x1480 + key * sizeof (hx_pointer)));
   }
   else if (key < 1088)
   {
@@ -232,7 +232,7 @@ hoox_tls_key_set_value (HooxTlsKey key,
 {
   if (key < 64)
   {
-    __writegsqword (0x1480 + key * sizeof (hx_pointer), (hx_uint64) value);
+    __writegsqword ((DWORD) (0x1480 + key * sizeof (hx_pointer)), (hx_uint64) value);
   }
   else if (key < 1088)
   {
@@ -246,7 +246,7 @@ hoox_tls_key_set_value (HooxTlsKey key,
     else
     {
       hoox_tls_key_set_tmp_value (key, value);
-      TlsSetValue (key, value);
+      TlsSetValue ((DWORD) key, value);
       hoox_tls_key_del_tmp_value (key);
     }
   }
